PCSX2-QT 1.7 64bit nightly support
PCSX2 is now 64bit only and it's base memory address for games is no longer static. However the devs
did export out a symbol for external programs to be able to locate the base pointer which is pcsx2-qt.EEmem.
Knowing that we can scan through the modules looking for that export to get the base pointer.