Fix tracker crash after a completed scenario and disable crown display in Single Player
We don't actually self correct for LUA starting at 1 index since the first room
is an initialization room so we just run with it as the first real room of a scenario
starts at 1. We had a check to hide the crown if the room was that initialization room.
However it would crash beforehand as we didn't have a 0 index since LUA starts at 1.
Moved the room check above so that the later code doesn't run when in the init room.